requirements/dev.txt
-r base.txt
-r test.txt

ipdb==0.11
ipython==6.4.0
flask-shell-ipython==0.3.1